% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_connect}
\alias{aou_connect}
\title{Connect to big query database in All of Us}
\usage{
aou_connect(bucket_name = "bucket")
}
\arguments{
\item{bucket_name}{variable name for your bucket. Recommend leaving the default.}
}
\description{
Connect to big query database in All of Us
}
