% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_get_survey_concepts}
\alias{aou_get_survey_concepts}
\title{Get survey questions from AoU for a given cohort}
\usage{
aou_get_survey_concepts(cohort, concepts, combine = FALSE, ...)
}
\arguments{
\item{cohort}{tbl; reference to a table with a column called "person_id"}

\item{concepts}{num; a vector of concept ids for questions in the survey table}

\item{combine}{lgl; whether to combine the question and answer into a single variable (used when pulling concepts across domains)}
}
\value{
a remote tbl with columns person_id, date (survey_datetime),
 concept_id (question_concept_id), question, answer. If combine = TRUE, concept_name (paste0(question, ":::", answer)), domain (= "Survey")
}
\description{
Get survey questions from AoU for a given cohort
}
\examples{
\dontrun{
survey_data <- aou_get_survey_concepts(cohort, concepts = c(1157, 124839))
 )}

}
