% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_ls}
\alias{aou_ls}
\title{List the current files in your bucket.}
\usage{
aou_ls(pattern = "*.csv")
}
\arguments{
\item{pattern}{pattern like *.csv or a single file name e.g., mydata.csv}

\item{description}{Quick function to list files matching a pattern in your bucket}
}
\description{
List the current files in your bucket.
}
