% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_ls_bucket}
\alias{aou_ls_bucket}
\title{List the current files in your bucket.}
\usage{
aou_ls_bucket(pattern = "*.csv", bucket_name = Sys.getenv("WORKSPACE_BUCKET"))
}
\arguments{
\item{pattern}{pattern like *.csv or a single file name e.g., mydata.csv}

\item{bucket_name}{name of your bucket. Recommend leaving the default}

\item{description}{Quick function to list files matching a pattern in your bucket}
}
\description{
List the current files in your bucket.
}
