% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_ls_workspace}
\alias{aou_ls_workspace}
\title{List the current files in your workspace.}
\usage{
aou_ls_workspace(pattern = "*.csv")
}
\arguments{
\item{pattern}{pattern like *.csv or a single file name e.g., mydata.csv}

\item{description}{Quick function to list files matching a pattern in your workspace}
}
\description{
List the current files in your workspace.
}
