% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_retrieve_from_bucket}
\alias{aou_retrieve_from_bucket}
\title{Retrieve a file from a bucket to your workspace}
\usage{
aou_retrieve_from_bucket(file_name, bucket_name = "bucket")
}
\arguments{
\item{file_name}{The name of a file in your bucket or a vector of multiple files.}

\item{bucket_name}{Name of your bucket. Recommend leaving the default}

\item{description}{This step retrieves a file you have saved permanently in your bucket
into your  where you can read it into R using a function like write.csv().}
}
\description{
Retrieve a file from a bucket to your workspace
}
