% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU.R
\name{aou_save_to_bucket}
\alias{aou_save_to_bucket}
\title{Save a file from your workspace to your bucket.}
\usage{
aou_save_to_bucket(file_name, bucket_name = "bucket")
}
\arguments{
\item{file_name}{name of file to save}

\item{bucket_name}{name of your bucket. Recommend leaving the default}

\item{description}{This step permanently saves a file you have saved in your workspace
to your bucket where you can always retrieve it. To use, first you need to save the desired
r object as a file (e.g., write.csv(object, filename.csv)) and then run this function
(e.g., aou_save_to_bucket(file_name = "filename.csv")).}
}
\description{
Save a file from your workspace to your bucket.
}
