% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt2omop.R
\name{cpt2omop}
\alias{cpt2omop}
\title{cpt2omop}
\usage{
cpt2omop(
  db_con,
  codes,
  cdm_schema = NULL,
  collect = TRUE,
  translate_from = "CPT4"
)
}
\arguments{
\item{db_con}{database connection object}

\item{codes}{vector of CPT4 codes}

\item{cdm_schema}{name of CDM schema}

\item{collect}{whether to return a dataframe (default) or sql query (Set to FALSE)}
}
\value{
a dataframe of icd, SNOMED, and OMOP concept codes
}
\description{
translates cpt codes to omop concept ID
}
