% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_collect.R
\name{dbi_collect}
\alias{dbi_collect}
\title{Collect using the DBI package}
\usage{
dbi_collect(query, connection = getOption("con.default.value"))
}
\arguments{
\item{query}{the sql query built from dbplyr}

\item{connection}{connection to the db. can leave blank if set using options}
}
\value{
collected dataframe
}
\description{
Collect using the DBI package
}
