% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd2omop2.R
\name{icd2omop2}
\alias{icd2omop2}
\title{icd2omop2}
\usage{
icd2omop2(
  db_con,
  codes,
  cdm_schema = NULL,
  icd_wildcard = "x",
  dbms_wildcard = "\%",
  translate_from = "ICD9CM",
  translate_to = "SNOMED"
)
}
\arguments{
\item{db_con}{database connection object}

\item{codes}{vector of ICD codes with or without wildcards}

\item{cdm_schema}{name of CDM schema}

\item{icd_wildcard}{wordcard indicator in icd codes. for example if code is R47.x, put "x"}

\item{dbms_wildcard}{wildcard indicator for dbms SIMILAR TO function.}

\item{translate_from}{ICD codes to start}

\item{translate_to}{codes to transfer to to get to omop. usually SNOMED}
}
\value{
a dataframe of icd, SNOMED, and OMOP concept codes
}
\description{
this function will only find exact ICD matches. It is faster than
icd2omop() because it does not need to use a join to find wildcards. It also does not
return a match with the beginning icd codes however. It is important to check that
all codes searched have found a match
}
