% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertTable_chunks.R
\name{insertTable_chunk}
\alias{insertTable_chunk}
\title{Insert a table to your user schema in chunks because speed and bugs}
\usage{
insertTable_chunk(
  data,
  table_name,
  n = 100,
  overwrite = TRUE,
  con = getOption("con.default.value"),
  write_schema = getOption("write_schema.default.value"),
  ...
)
}
\arguments{
\item{data}{data you want to write to your user schema}

\item{table_name}{the name of the table that you want to write to the database}

\item{n}{the number of chunks to split. defaults to 100. Aim for < 200 rows per chunk right now.}

\item{overwrite}{do you want to overwrite an existing table of the same name?}

\item{con}{connection. defaults to the set option if done.}

\item{...}{additional arguments to be passed along to DatabaseConnector::insertTable}

\item{user_schema}{your user schema. defaults to the set option if done}
}
\description{
Insert a table to your user schema in chunks because speed and bugs
}
\examples{
\dontrun{
options(con.default.value = con)
write_schema = paste0("work_", keyring::key_get("db_username"))
insertTable_chunk(data = data, table_name = "table1", n = 50, overwrite = TRUE, user_schema = write_schema)
}
}
