% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt2omop.R
\name{map2omop}
\alias{map2omop}
\title{map2omop}
\usage{
map2omop(
  db_con,
  codes,
  cdm_schema = NULL,
  collect = TRUE,
  translate_from = "CPT4"
)
}
\arguments{
\item{db_con}{database connection object}

\item{codes}{vector of CPT4 codes}

\item{cdm_schema}{name of CDM schema}

\item{collect}{whether to return a dataframe (default) or sql query (Set to FALSE)}

\item{translate_from}{vocab to translate from. Can be a vector e.g., c("CPT4", "HCPCS")}
}
\value{
a dataframe of icd, SNOMED, and OMOP concept codes
}
\description{
translates any codes to omop concept ID as long as there is a direct
mapping in the concept table
}
