% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_health_icd.R
\name{mat_health_icd}
\alias{mat_health_icd}
\title{Maternal Health ICD replication codes}
\usage{
mat_health_icd()
}
\value{
dataframe of icd codes
}
\description{
Maternal Health ICD replication codes
}
\examples{
mat_health_icd()
}
