% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsilab_connect.R
\name{ohdsilab_connect}
\alias{ohdsilab_connect}
\title{Connect to the ohdsilab redshift database}
\usage{
ohdsilab_connect(username, password, cdm_schema = "omop_cdm_53_pmtx_202203")
}
\arguments{
\item{username}{your redshift username (required)}

\item{password}{your redshift password (required)}

\item{cdm_schema}{the omop cdm schema (set by default)}
}
\value{
a connection object
}
\description{
Helper function to connect to the ohdsilab OMOP database. You will need to
supply your redshift username and password (which you received in an email when
you first set up your workspace). The location of the CDM schema is set by default.

The function will also set default values for your user schema, the cdm schema,
and the connection object, which are used by other functions in the ohdsilab package.
These can be found by running getOption("con.default.value"), getOption("schema.default.value"),
and getOption("write_schema.default.value").
}
\examples{

\dontrun{
con <- ohdsilab_connect(
    username = keyring::key_get("lab_user"),
    password = keyring::key_get("lab_password")
    )
}
}
