% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_join.R
\name{omop_join}
\alias{omop_join}
\title{Join current query to another omop table}
\usage{
omop_join(
  data,
  table,
  type,
  con = getOption("con.default.value"),
  schema = getOption("schema.default.value"),
  ...
)
}
\arguments{
\item{data}{sql query from dbplyr/dplyr. this function works in pipes!}

\item{table}{the omop table (or other table in your schema) you wish to join}

\item{type}{the type of join. use types available in dplyr: left, right, inner, anti, full etc.}

\item{con}{defaults to the connection you set with options()}

\item{schema}{defaults to the schema you set with options()}

\item{...}{arguments passed on to the join function. e.g., by = "person_id"}
}
\value{
more sql query info stuff
}
\description{
Simple wrapper for join functions to join an existing query
to another table in the omop cdm (or any other of the same source). Include
the following two lines at the top of your script after setting the schema and connection

\code{options(con.default.value = connection_variable_name)}

\code{options(schema.default.value = cdm_schema_variable_name)}
}
\examples{
options(con.default.value = con)
options(schema.default.value = cdm_schema)
obs_tbl |>
  omop_join("person", type = "left", by = "person_id")

}
