% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollback.R
\name{rb}
\alias{rb}
\title{shortcut to rollback in teh case of a redshift error}
\usage{
rb(con = getOption("con.default.value"))
}
\arguments{
\item{con}{defaults to the default connection in the session. either set using options or supplied directly}
}
\value{
nothing
}
\description{
shortcut to rollback in teh case of a redshift error
}
