% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedQuery.R
\name{seedQuery}
\alias{seedQuery}
\title{Execute a query with a seed for reproducibility}
\usage{
seedQuery(query, seed, con = getOption("con.default.value"))
}
\arguments{
\item{query}{SQL query generated by dplyr}

\item{seed}{A seed you'd like to set - a value between 0 and 1, inclusive}

\item{con}{the connection to the redshift server for ohdsilab.}
}
\description{
Execute a query with a seed for reproducibility
}
