% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedQuery.R
\name{set_seed}
\alias{set_seed}
\title{Set a seed for reproducibility}
\usage{
set_seed(seed, con = getOption("con.default.value"))
}
\arguments{
\item{seed}{A seed you'd like to set - a value between 0 and 1, inclusive}

\item{con}{the connection to the redshift server for ohdsilab.}
}
\description{
Set a seed for reproducibility
}
