% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_insert 2.R, R/str_insert.R
\name{str_insert}
\alias{str_insert}
\title{Insert string}
\usage{
str_insert(x, pos, insert)

str_insert(x, pos, insert)
}
\arguments{
\item{x}{a character or character vector}

\item{pos}{what position in the string to insert}

\item{insert}{what to insert}
}
\value{
icd code with a period
}
\description{
Useful for inserting periods in icd9 codes when they come without.

Useful for inserting periods in icd9 codes when they come without.
}
