% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_database.R
\name{write_to_database}
\alias{write_to_database}
\title{Write the result of a query to your user Schema}
\usage{
write_to_database(
  input_tbl,
  con = getOption("con.default.value"),
  schema,
  table_name,
  db = "ohdsi_lab"
)
}
\arguments{
\item{input_tbl}{The sql query generated by dbplyr/dplyr}

\item{con}{the database connection, defaults to what is set by options()}

\item{schema}{your user schema}

\item{table_name}{The name of the new table you would like to create}

\item{db}{The name of the database, defaults to ohdsi_lab}
}
\description{
Write the result of a query to your user Schema
}
