% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd2omop.R
\name{icd2omop}
\alias{icd2omop}
\title{icd2omop}
\usage{
icd2omop(
  db_con,
  schema = FALSE,
  codes,
  schema.name = NA,
  icd_wildcard = "x",
  dbms_wildcard = "\%",
  icd_version = "ICD9CM"
)
}
\arguments{
\item{db_con}{database connection object}

\item{schema}{if there is a database schema, TRUE. defaults of FALSE}

\item{codes}{vector of ICD codes with or without wildcards}

\item{schema.name}{name of databsae schema. only required if schema is TRUE}

\item{icd_wildcard}{wordcard indicator in icd codes. for example if code is R47.x, put "x"}

\item{dbms_wildcard}{wildcard indicator for dbms SIMILAR TO function.}

\item{icd_version}{ICD9CM or ICD10CM. Make sure to enter the same way it is included in the concept table}
}
\value{
a dataframe of icd, SNOMED, and OMOP concept codes
}
\description{
icd2omop
}
