% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_condition_concepts.R
\name{pull_concept_set}
\alias{pull_concept_set}
\title{Get occcurances of of a concept set from ATLAS for a given cohort}
\usage{
pull_concept_set(
  cohort,
  concept_set_id,
  start_date,
  end_date,
  name = "has_concept",
  min_n = NULL,
  n = FALSE,
  keep_all = FALSE,
  con,
  base_url,
  my_schema
)
}
\arguments{
\item{cohort}{reference to a table created in a user schema with a column called "person_id", and columns for start_date and end_date}

\item{concept_set_id}{num; the number associated with the concept_set_id in ATLAS}

\item{start_date}{the name of the start_date column in the cohort table (unquoted)}

\item{end_date}{the name of the end_date column in the cohort table (unquoted)}

\item{name}{Name of column for indicator variable returned by function}

\item{min_n}{the minimum number of occurrences per person to consider the indicator true}

\item{n}{count the number of occurrences per person (will not include zeros)}

\item{keep_all}{keep columns with information about the concept (e.g., concept name, id, etc.)}

\item{con}{the connection object to ohdsilab}

\item{base_url}{the base url for ATLAS (not ohdsilab)}

\item{my_schema}{schema for the user}
}
\value{
a dataframe
}
\description{
Get occcurances of of a concept set from ATLAS for a given cohort
}
\examples{
\dontrun{
tobacco <- pull_concepts(con, cohort, concept_set_id = 1157, start_date = covariate_start_date,
 end_date = cohort_start_date, base_url = base_url, name = "tobacco", my_schema = my_schema
 )}

}
